// wetty-chat-mobile/src/generated/models/serverWsMessageType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wetty Chat API
 * Real-time chat application backend API supporting groups, messaging, threads, stickers, invites, and push notifications.
 * OpenAPI spec version: 0.1.0
 */

export enum ServerWsMessageType {
  message = 'message',
  messageUpdated = 'messageUpdated',
  messageDeleted = 'messageDeleted',
  messagesBulkDeleted = 'messagesBulkDeleted',
  reactionUpdated = 'reactionUpdated',
  presenceUpdate = 'presenceUpdate',
  threadUpdate = 'threadUpdate',
  threadMembershipChanged = 'threadMembershipChanged',
  chatArchiveStateChanged = 'chatArchiveStateChanged',
  pinAdded = 'pinAdded',
  threadPinAdded = 'threadPinAdded',
  pinRemoved = 'pinRemoved',
  threadPinRemoved = 'threadPinRemoved',
  stickerPackOrderUpdated = 'stickerPackOrderUpdated',
  friendRequestReceived = 'friendRequestReceived',
  friendRequestResolved = 'friendRequestResolved',
  friendshipRemoved = 'friendshipRemoved',
}